requirements.txt
torch
torchvision
torchaudio
diffusers==0.36.0
transformers==4.55.2
accelerate
einops
easydict
numpy==1.26.4
tqdm
imageio[ffmpeg]
websockets
msgpack
opencv-python
matplotlib
ftfy
safetensors
Pillow

# Post-training
# lerobot==0.3.3
scipy
wandb
